.history/requirements_20250830135125.txt
Django==4.2.19
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
pymongo==4.6.1
xtquant==1.0.0
PyJWT==2.8.0
cryptography==41.0.7 